Move union-find to own file
